+++ src/main/java/com/spring/practice/restjpa/repository/EmployeeRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface EmployeeRepository
-        extends JpaRepository<Employee, Integer> {
+        extends JpaRepository<Employee, Long> {
 
     List<Employee> findAllByName(String name);
 }
